More quality fixes. (#160)
* Add a custom diagnostic ID for the obsolete enum names.

* Add a custom diagnostic ID for obsolete data types.

* Give clearer names to some `FileSystemType` names.

`Hdfs` was not renamed; the acronym is more established. On the other hand, the top Google Search results for "GCS" point to the Glascow Coma Scale.

* Obsolete the `TileDB.Interop` namespace, suppressing the warning for our projects.

We can't have partial enums so I modified the auto-generated enum files; it will be temporary.

* Reformat the obsoletions document.

* Fix badly formed XML documentation.

* Mitigate a breaking change in the `File` class.

* Add diagnostic IDs to obsolete methods of the `EnumUtil` class.

* Remove the `nuint` `File.BufferImport`/`BufferExport` overloads.

Let's stick with just `ulong`, there might be ambiguities if there are both `ulong` and `nuint`.