Refactor the buffer setter methods to be more resilient against exceptions.
We used to first register the buffer in the query's dictionary and then in the Core, but this approach was wrong and results in memory leaks or worse in case the native code throws an exception.
The correct way is to pin the new buffer, register it in the Core, and then create or update the dictionary entry.
We also stopped reusing the `BufferHandle` objects for simplicity (and maybe correctness?).