Fix nightlies et al. (#487)
* Remove "Remove existing .NET versions" CI step.
This step is not necessary because the .NET SDK version to use is pinned by the `global.json` file.
Moreover it's likely that the forced deletion of the `$DOTNET_ROOT` folder corrupts the installation and causes nightlies to fail because the `dotnet --info` command fails with an exception.

* Update to the latest feature band of the .NET 8 SDK.

And update the setup-dotnet steps to install the latest .NET SDK patch.

* Update MSTest packages together.